created a new package 'fml', moved fml code to package and updated scikit build
