+++ sklearn/fml/__init__.py
@@ -1,0 +1,7 @@
+"""
+The :mod:`sklearn.fml` module implements Federated Meta Learning
+"""
+
+from .fml_client import FMLClient
+
+__all__ = ['FMLClient']
